requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Type hints support
typing-extensions==4.8.0

# Configuration management
python-dotenv==1.0.0

# LangExtract and dependencies
langextract==0.1.10
google-generativeai==0.3.2

# PDF processing
PyPDF2==3.0.1
pdfplumber==0.10.3
pdf2image==1.17.0
Pillow==10.1.0
python-magic==0.4.27

# OCR support
pytesseract==0.3.10
opencv-python==4.8.1.78

# Text processing
spacy==3.7.2
nltk==3.8.1

# Background tasks and caching
celery[redis]==5.3.4
redis==5.0.1
aioredis==2.0.1

# Database (for caching and job persistence)
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# File handling and utilities
aiofiles==23.2.1
python-slugify==8.0.1
mmh3==4.0.1

# Monitoring and logging
structlog==23.2.0
prometheus-client==0.19.0

# Rate limiting
slowapi==0.1.9

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0